moved content storge into its own class
